Share file-node scenario coverage across backends
Parameterize the root-like `prepare_path` unit test cases and move the
backend-neutral file-node ACL/path scenarios into a shared test module so both
SQLite and Postgres execute them.

Add SQLite legacy default-deny checks before legacy ACL rows are inserted, and
redact credentials from Postgres teardown connection diagnostics.
